Renamed ActivityAiButton to CameraButton.
