+++ components/Camera/CameraButton.tsx
